받아오기

+++ src/main/java/com/tptp/controller/OController.java
@@ -24,6 +24,36 @@ public class OController {
 	@Resource(name = "oService")
 	private OService oService;
 
+	@RequestMapping(value = "ohSearch.do")
+	public ModelAndView ohSearch(HttpServletRequest request, CommandMap commandMap) throws Exception {
+		ModelAndView mv = new ModelAndView();
+		
+		System.out.println(request.getParameter("b_cate1"));
+		if (!request.getParameter("searchCont").equals("")) {
+			List<Map<String, Object>> list = oService.ohSearch(commandMap.getMap());
+			if(list.size() > 0) {			
+				mv.addObject("resultSearch", list);
+			}
+			System.out.println(list);
+			System.out.println(commandMap.get("b_cate1"));
+			
+			if(commandMap.get("b_cate1").equals("b1")) {
+				System.out.println("b1으로 옴");
+				mv.setViewName("brand1");				
+			}else if(commandMap.get("b_cate1").equals("b2")) {
+				System.out.println("b2으로 옴");
+				mv.setViewName("brand2");
+			}else if(commandMap.get("b_cate1").equals("b3")) {
+				System.out.println("b3으로 옴");
+				mv.setViewName("brand3");
+			}else if(commandMap.get("b_cate1").equals("no")) {
+				System.out.println("no로 옴");
+				mv.setViewName("notice");
+			}
+		}
+		return mv;
+	}
+	
 	@RequestMapping(value = "notice.do")
 	public ModelAndView notice(HttpServletRequest request, CommandMap commandMap) throws Exception {
 		ModelAndView mv = new ModelAndView();
@@ -31,27 +61,25 @@ public class OController {
 		HttpSession session = request.getSession();
 		
 		// 임시로 page 만들어주기
-		int page = 1;
-		if (commandMap.containsKey("page") && Integer.parseInt((String) commandMap.get("page")) > 0) {
-			page = Integer.parseInt((String) commandMap.get("page"));
-		}
-		
-		commandMap.put("page", (page - 1) * 10);// '0'
-		List<Map<String, Object>> notice = oService.notice(commandMap.getMap());
-		
-		mv.addObject("notice", notice);
-		mv.addObject("page", page);
+				int page = 1;
+				if (commandMap.containsKey("page") && Integer.parseInt((String) commandMap.get("page")) > 0) {
+					page = Integer.parseInt((String) commandMap.get("page"));
+				}
+				commandMap.put("page", (page - 1) * 10);// '0'
 
-		mv.addObject("b_cate1", notice.get(0).get("b_cate1"));
-		System.out.println(commandMap.getMap());
+				List<Map<String, Object>> notice = oService.brand1(commandMap.getMap());
+				
+				if(notice.size() > 0) {
+					
+					mv.addObject("notice", notice);
+			
+					mv.addObject("b_cate1", notice.get(0).get("b_cate1"));
+					mv.addObject("b_cate2", commandMap.get("b_cate2"));
+					
+					mv.addObject("page", page);
+					mv.addObject("count", notice.get(0).get("count"));
 
-
-		if (notice.size() > 0) {
-			// 총페이지는?
-			// System.out.println("총 글의 수 : " + list.get(0).get("COUNT"));
-			commandMap.put("l_id", session.getAttribute("id"));
-			mv.addObject("count", notice.get(0).get("count"));
-		}
+				}
 		return mv;
 	}
 
@@ -75,13 +103,33 @@ public class OController {
 			mv.addObject("brand1", brand1);
 	
 			mv.addObject("b_cate1", brand1.get(0).get("b_cate1"));
-			mv.addObject("b_cate2", brand1.get(0).get("b_cate2"));
+			mv.addObject("b_cate2", commandMap.get("b_cate2"));
 			
 			mv.addObject("page", page);
+			if (commandMap.get("b_cate2") != null) {
+				if (commandMap.get("b_cate2").equals("연필")) {
+					mv.addObject("count", brand1.get(0).get("pen_count"));
+					
+				}else if (commandMap.get("b_cate2").equals("볼펜")) {
+					mv.addObject("count", brand1.get(0).get("ball_count"));
 			
-			mv.addObject("count", brand1.get(0).get("count"));
+				}else if (commandMap.get("b_cate2").equals("샤프")) {
+					mv.addObject("count", brand1.get(0).get("sha_count"));
+					
+				}else if (commandMap.get("b_cate2").equals("형광펜")) {
+					mv.addObject("count", brand1.get(0).get("hi_count"));
+					
+				}else if (commandMap.get("b_cate2").equals("만년필")) {
+					mv.addObject("count", brand1.get(0).get("foun_count"));
+					
+				}else if (commandMap.get("b_cate2").equals("기타")) {
+					mv.addObject("count", brand1.get(0).get("etc_count"));
+				}
+			}else if(commandMap.get("b_cate2") == null) {
+				mv.addObject("count", brand1.get(0).get("count"));
+				System.out.println(brand1.get(0).get("count"));
+			}
 			
-			System.out.println(brand1.get(0).get("count"));
 		}
 
 		return mv;
@@ -92,7 +140,7 @@ public class OController {
 		ModelAndView mv = new ModelAndView();
 		
 		HttpSession session = request.getSession();
-
+		
 		// 임시로 page 만들어주기
 		int page = 1;
 		if (commandMap.containsKey("page") && Integer.parseInt((String) commandMap.get("page")) > 0) {
@@ -101,17 +149,38 @@ public class OController {
 		commandMap.put("page", (page - 1) * 10);// '0'
 
 		List<Map<String, Object>> brand2 = oService.brand2(commandMap.getMap());
-
-		mv.addObject("brand2", brand2);
-
-		mv.addObject("b_cate1", brand2.get(0).get("b_cate1"));
-
-		mv.addObject("page", page);
-
-		if (brand2.size() > 0) {
-			// 총페이지는?
-			// System.out.println("총 글의 수 : " + list.get(0).get("COUNT"));
-			mv.addObject("count", brand2.get(0).get("count"));
+		
+		if(brand2.size() > 0) {
+			
+			mv.addObject("brand2", brand2);
+	
+			mv.addObject("b_cate1", brand2.get(0).get("b_cate1"));
+			mv.addObject("b_cate2", brand2.get(0).get("b_cate2"));
+			
+			mv.addObject("page", page);
+			if (commandMap.get("b_cate2") != null) {
+				if (commandMap.get("b_cate2").equals("연필")) {
+					mv.addObject("count", brand2.get(0).get("pen_count"));
+					
+				}else if (commandMap.get("b_cate2").equals("볼펜")) {
+					mv.addObject("count", brand2.get(0).get("ball_count"));
+			
+				}else if (commandMap.get("b_cate2").equals("샤프")) {
+					mv.addObject("count", brand2.get(0).get("sha_count"));
+					
+				}else if (commandMap.get("b_cate2").equals("형광펜")) {
+					mv.addObject("count", brand2.get(0).get("hi_count"));
+					
+				}else if (commandMap.get("b_cate2").equals("만년필")) {
+					mv.addObject("count", brand2.get(0).get("foun_count"));
+					
+				}else if (commandMap.get("b_cate2").equals("기타")) {
+					mv.addObject("count", brand2.get(0).get("etc_count"));
+				}
+			}else if(commandMap.get("b_cate2") == null) {
+				mv.addObject("count", brand2.get(0).get("count"));	
+			}
+			
 		}
 
 		return mv;
@@ -122,7 +191,7 @@ public class OController {
 		ModelAndView mv = new ModelAndView();
 		
 		HttpSession session = request.getSession();
-
+		
 		// 임시로 page 만들어주기
 		int page = 1;
 		if (commandMap.containsKey("page") && Integer.parseInt((String) commandMap.get("page")) > 0) {
@@ -131,17 +200,38 @@ public class OController {
 		commandMap.put("page", (page - 1) * 10);// '0'
 
 		List<Map<String, Object>> brand3 = oService.brand3(commandMap.getMap());
+		
+		if(brand3.size() > 0) {
+			
+			mv.addObject("brand3", brand3);
+	
+			mv.addObject("b_cate1", brand3.get(0).get("b_cate1"));
+			mv.addObject("b_cate2", brand3.get(0).get("b_cate2"));
+			
+			mv.addObject("page", page);
+			if (commandMap.get("b_cate2") != null) {
+				if (commandMap.get("b_cate2").equals("연필")) {
+					mv.addObject("count", brand3.get(0).get("pen_count"));
+					
+				}else if (commandMap.get("b_cate2").equals("볼펜")) {
+					mv.addObject("count", brand3.get(0).get("ball_count"));
+			
+				}else if (commandMap.get("b_cate2").equals("샤프")) {
+					mv.addObject("count", brand3.get(0).get("sha_count"));
+					
+				}else if (commandMap.get("b_cate2").equals("형광펜")) {
+					mv.addObject("count", brand3.get(0).get("hi_count"));
+					
+				}else if (commandMap.get("b_cate2").equals("만년필")) {
+					mv.addObject("count", brand3.get(0).get("foun_count"));
+					
+				}else if (commandMap.get("b_cate2").equals("기타")) {
+					mv.addObject("count", brand3.get(0).get("etc_count"));
+				}
+			}else if(commandMap.get("b_cate2") == null) {
+				mv.addObject("count", brand3.get(0).get("count"));	
+			}
 
-		mv.addObject("brand3", brand3);
-
-		mv.addObject("b_cate1", brand3.get(0).get("b_cate1"));
-
-		mv.addObject("page", page);
-
-		if (brand3.size() > 0) {
-			// 총페이지는?
-			// System.out.println("총 글의 수 : " + list.get(0).get("COUNT"));
-			mv.addObject("count", brand3.get(0).get("count"));
 		}
 
 		return mv;
@@ -155,6 +245,11 @@ public class OController {
 		
 		List<Map<String, Object>> commdetail = oService.commdetail(commandMap.getMap());
 		Map<String, Object> detail = oService.detail(commandMap.getMap());
+		
+		String old_url = request.getHeader("Referer");
+		String old_url2 = old_url.substring(26);
+		mv.addObject("old_url2", old_url2);
+		System.out.println(old_url2);
 		
 		// 조회수 + 1
 		int num = Integer.parseInt(request.getParameter("b_no"));
@@ -178,10 +273,14 @@ public class OController {
 		
 		HttpSession session = request.getSession();
 		
+		String old_url = request.getHeader("Referer");
+		String old_url2 = old_url.substring(26);
+		mv.addObject("old_url2", old_url2);
+		
 		if (session.getAttribute("id") != null) {
 			mv.addObject("b_cate1", commandMap.get("b_cate1"));			
 		} else {
-			mv.setViewName("redirect:brand1.do");
+			mv.setViewName("redirect:" + old_url2);
 		}
 
 		return mv;
@@ -193,18 +292,12 @@ public class OController {
 		ModelAndView mv = new ModelAndView();
 		
 		HttpSession session = request.getSession();
-
-		if (request.getParameter("b_title") != null && request.getParameter("b_content") != null) {
-
-			System.out.println(session.getAttribute("nick"));
-			
+		
+		if (request.getParameter("b_title") != null && request.getParameter("b_content") != null) {			
 			commandMap.put("l_nick", session.getAttribute("nick"));
-			System.out.println(commandMap.getMap());
 			int result = oService.writeInsert(commandMap.getMap());
 
-//			String old_url = request.getHeader("Referer");
-//			System.out.println(old_url);
-			mv.setViewName("redirect:brand1.do");
+			mv.setViewName("redirect:" + commandMap.get("old_url2"));
 		} else {
 			mv.setViewName("redirect:write.do");
 		}
@@ -218,20 +311,15 @@ public class OController {
 		
 		HttpSession session = request.getSession();
 		
-//		System.out.println(request.getParameter("url"));
-//		String url = request.getParameter("url").substring(17);
-//		String url2 = url.replace(".jsp", ".do");
-//		String old_url = request.getHeader("Referer");
-//		String old_url2 = old_url.substring(26);
-//		System.out.println(url2);
-		int delete = oService.delete(commandMap.getMap());
-		
-		if (delete == 1) {
-//			mv.setViewName("redirect:" + url2);
-			mv.setViewName("redirect:brand1.do");
+		if(session.getAttribute("id") != null) {
+			int delete = oService.delete(commandMap.getMap());
+			if (delete == 1) {
+				mv.setViewName("redirect:" + commandMap.get("old_url2"));
+			} else {
+				mv.setViewName("redirect:detail.do?b_no" + commandMap.get("b_no"));
+			}
 		}
-		mv.addObject("delete", delete);
-	
+		
 		return mv;
 	}
 	
@@ -242,13 +330,21 @@ public class OController {
 		HttpSession session = request.getSession();
 		
 		Map<String, Object> update = oService.update(commandMap.getMap());
-		mv.addObject("update", update);
 		
+		
+		String old_url = request.getHeader("Referer");
+		String old_url2 = old_url.substring(26);
+		mv.addObject("old_url2", old_url2);
 		
 		System.out.println(commandMap.getMap());
 		System.out.println(request.getParameter("b_title"));
 		System.out.println(request.getParameter("b_content"));
 		
+		if (session.getAttribute("id") != null) {
+			mv.addObject("update", update);
+		} else {
+			mv.setViewName("redirect:" + old_url2);
+		}
 		
 		
 //		mv.addObject("b_cate1", commandMap.get("b_cate1"));
@@ -264,11 +360,16 @@ public class OController {
 		
 		HttpSession session = request.getSession();
 		
-		int result = oService.updateInsert(commandMap.getMap());
-		System.out.println("result : " + result);
-		System.out.println("map : " + commandMap.getMap());
-		
-		mv.setViewName("redirect:brand1.do");
+		if (request.getParameter("b_title") != null && request.getParameter("b_content") != null) {			
+			commandMap.put("l_nick", session.getAttribute("nick"));
+			int result = oService.updateInsert(commandMap.getMap());
+			System.out.println("result : " + result);
+			System.out.println("map : " + commandMap.getMap());
+
+			mv.setViewName("redirect:" + commandMap.get("old_url2"));
+		} else {
+			mv.setViewName("redirect:update.do");
+		}
 		
 		return mv;
 	}
@@ -278,10 +379,6 @@ public class OController {
 		ModelAndView mv = new ModelAndView();
 		
 		HttpSession session = request.getSession();
-		
-		int result = oService.updateInsert(commandMap.getMap());
-		System.out.println("result : " + result);
-		System.out.println("map : " + commandMap.getMap());
 		
 		
 		return mv;
